feat: kiosk variant picker, KDS wiring, loyalty cleanup, platform banners & QR card layouts
Squashes WIP commits; drops marketing/.mcp.json from tracking (now gitignored, contained an API key).

- Kiosk self-order: variant picker (bottom-sheet, per-variant cart, variant_id)
- KDS toggle now actually gates kitchen routing + hides board/sidebar when off
- Customers: remove dead order-count segments; keep Blacklist; remove dead /store/segments endpoint
- Membership tier shown on customer detail (single loyalty system)
- Platform banners: admin CRUD + dashboard slider + drag-reorder
- QR table cards: layout templates (classic/tent/poster) + colors/headline/caption, print + queue-board link
- Misc UI fixes (mobile banner overflow, blog cover fit, admin topbar, favicon)

+++ api/internal/repository/bulk_jobs.go
@@ -106,7 +106,7 @@ func (r *BulkJobRepo) Complete(ctx context.Context, id uuid.UUID, succeeded, fai
 	_, err = r.pool.Exec(ctx, `
 		UPDATE bulk_jobs
 		SET status = 'completed',
-		    processed_rows = $2 + $3,
+		    processed_rows = $2::int + $3::int,
 		    succeeded = $2,
 		    failed = $3,
 		    errors_json = $4::jsonb,
